+++ Plugins/SqlEnterpriseFormatter/formatorderby.cpp
@@ -12,4 +12,7 @@ void FormatOrderBy::formatInternal()
     withStatement(orderBy->expr);
     if (orderBy->order != SqliteSortOrder::null)
         withKeyword(sqliteSortOrder(orderBy->order));
+
+    if (orderBy->nulls != SqliteNulls::null)
+        withKeyword("NULLS").withKeyword(sqliteNulls(orderBy->nulls));
 }